Stop the service if it isn't needed
Firstly, a service is needed because HEADSET_PLUG events are only
delivered to running processes, not <receiver>'s. This is caused
by a flag setting on HeadsetObserver.

However, the service can be stopped if it isn't strictly speaking
necessary. The condition used right now is that if no effect
checkboxes are enabled, the service will do nothing.

To accomplish automatic start/stop, I made every preference change
publish an UPDATE event which goes to the receiver, which checks the
configuration to evaluate if the service is needed.

The service, whenever it is running, ensures that all preferences
changes and headset plug/unplug events get published to AudioManager.