Make equalizer smarter and larger to improve use
Separate Complex and Biquad to their own files. Equalizer now takes twice the
vertical space in its dialog version. User's touch is now mapped to the closest
slider. The frequency response curve extends past the start and end of the
equalizer, and is rendered with less points to improve responsivity. The
slider control is rendered with bold red to make it stand out better.