Update DSPManager to Android 2.3 AudioEffect API
While DSPManager uses the sdk 9's APIs, it is nevertheless dependent on
libcyanogen-dsp, because it constructs a dynamic range compressor effect
via reflection.

Incidentally, the DRC effect is still broken and can not be used. It has
been designed to be added to individual audio sessions, and not to the output
mix. This is something DSPManager can not currently do.

Current version of DRC effect is also otherwise buggy and causes loud sound
distortion, even from very quiet audio sources. You have been duly warned.

Change-Id: I82b7be4044ce81a296e2226263d4b543b87957fe