Make DSPManager persistent service
This application must be running at all times, so that it can respond to
AudioEffect intents and manage audio sessions. Some users complained about
DSPManager not always managing their audio sessions, and I saw that
Android sometimes kills the DSPManager process and then ends up restarting
it mere seconds later. This is a likely cause for the problems, because
DSPManager will forget the IDs of the audio sessions in a restart.

This bugfix should be backported to every maintained CM release.

Change-Id: I79c6bb89453a171d45191163238c94abdc2226fe